Add `Array::try_from_iter`
Adds a fallible equivalent to `FromIterator` which returns
`TryFromIteratorError` if the number of items in the iterator does not
match the size of the array.
